2023.11.24

+++ scripts/install_pve.sh
@@ -1285,7 +1285,7 @@ if grep -q "vmbr0" "/etc/network/interfaces"; then
     _blue "vmbr0 already exists in /etc/network/interfaces"
     _blue "vmbr0 已存在在 /etc/network/interfaces"
 else
-    if [ -z "$ipv6_address" ] || [ -z "$ipv6_prefixlen" ] || [ -z "$ipv6_gateway" ]; then
+    if [ -z "$ipv6_address" ] || [ -z "$ipv6_prefixlen" ] || [ -z "$ipv6_gateway" ] && [ ! -f /usr/local/bin/pve_last_ipv6 ]; then
         cat <<EOF | sudo tee -a /etc/network/interfaces
 auto vmbr0
 iface vmbr0 inet static
@@ -1295,7 +1295,7 @@ iface vmbr0 inet static
     bridge_stp off
     bridge_fd 0
 EOF
-    elif [ -f "/usr/local/bin/iface_auto.txt" ]; then
+    elif [ -f "/usr/local/bin/iface_auto.txt" ] && [ ! -f /usr/local/bin/pve_last_ipv6 ]; then
         cat <<EOF | sudo tee -a /etc/network/interfaces
 auto vmbr0
 iface vmbr0 inet static
@@ -1308,11 +1308,9 @@ iface vmbr0 inet static
 iface vmbr0 inet6 auto
     bridge_ports $interface
 EOF
-    else
-        # 与${ipv6_gateway}比较是否相同
-        if [ -f /usr/local/bin/pve_last_ipv6 ]; then
-            last_ipv6=$(cat /usr/local/bin/pve_last_ipv6)
-            cat <<EOF | sudo tee -a /etc/network/interfaces
+    elif [ -f /usr/local/bin/pve_last_ipv6 ]; then
+        last_ipv6=$(cat /usr/local/bin/pve_last_ipv6)
+        cat <<EOF | sudo tee -a /etc/network/interfaces
 auto vmbr0
 iface vmbr0 inet static
     address $ipv4_address
@@ -1328,8 +1326,8 @@ iface vmbr0 inet6 static
 iface vmbr0 inet6 static
     address ${ipv6_address_without_last_segment}1/128
 EOF
-        else:
-            cat <<EOF | sudo tee -a /etc/network/interfaces
+    else
+        cat <<EOF | sudo tee -a /etc/network/interfaces
 auto vmbr0
 iface vmbr0 inet static
     address $ipv4_address
@@ -1342,7 +1340,6 @@ iface vmbr0 inet6 static
     address ${ipv6_address_without_last_segment}1/128
     gateway ${ipv6_gateway}
 EOF
-        fi
     fi
 fi
 chattr +i /etc/network/interfaces
